requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
numpy
requests
pandas
plotly
scikit-learn
statsmodels
